# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Stanks
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /v0
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  Bearer:
    description: JWT authentication
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /user/{email}:
    x-swagger-router-controller: user
    get:
      description: Get user information
      operationId: getUser
      parameters:
        - name: email
          in: path
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UserResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/MessageResponse"
      security:
        - Bearer: []
...
definitions:
  MessageResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  UserResponse:
    required:
      - id
      - email
      - username
      - first_name
      - last_name
      - dob
      - created_at
      - updated_at
    properties:
      id:
        type: number
        format: int32
      email:
        type: string
      username:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      dob:
        type: string
        format: date-time
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time